Filesystem abstraction layer pushed in between the implementations and FileStorage
A few tweaks to the file abstraction layer have been incorporated as well.
There currently is a very naive implementation of resumable upload with streams in there as it is extremely vulnerable for DOS attacks.
This is due to it having to retrieve the full file and writing it with some extra data back again to some unknown remote for each chunk that gets uploaded. Next commit fixes this.